PROBLEM: ffi-rzmq tests crash with malloc errors
SOLUTION: change memory layout of class LibZMQ::Message

For some unknown reason, the tests of ffi-rzmq crash when using
any libzmq version higher than 4.0.7, even though the number of
bytes of the current declaration is correct for 4.1.1+.

I changed the layout to correspond exactly what zmq.h
exposes to a C program. This fixed all the crashes of
the tests and also my programs running on 4.1.4.